core/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-phonenumber-field==8.0.0
django-private-storage==3.1.1
phonenumbers==8.13.46
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
